MaxLabProject/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
decorator==4.4.2
deep-translator==1.5.4
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
mysql-client==0.0.1
mysqlclient==2.0.3
numpy==1.21.1
Pillow
proglog==0.1.9
pydub==0.25.1
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tqdm==4.62.0
urllib3==1.26.6
django-environ==0.9.0